music_recommendation/music_recommendation/requirements.txt
fastapi
uvicorn
pandas
numpy==1.23.5
faiss-cpu==1.7.4
scikit-learn
pydantic
surprise
pickleshare